requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
mysql-connector-python==8.2.0
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
scikit-learn==1.3.2
Werkzeug==3.0.1